+++ game/graph/Edge.java
@@ -1,0 +1,10 @@
+package game.graph;
+
+public class Edge {
+    public Node a;
+    public Node b;
+    
+    public boolean linked(Node node) {
+        return a == node || b == node;
+    }
+}
