+++ game/graph/Node.java
@@ -1,0 +1,6 @@
+package game.graph;
+
+public class Node {
+    public Node[] myNodes;
+    public Edge[] myEdges;
+}
